First espresso tests
