+++ .gitignore
@@ -6,6 +6,3 @@ App/build
 *.iml
 .gradle
 build
-*.pyc
-*.py~
-venv*
