+++ App/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'com.android.application'
-apply plugin: 'robolectric'
 
 android {
     compileSdkVersion 21
@@ -13,7 +12,7 @@ android {
         targetSdkVersion 21
         versionCode 325
         versionName "3.2.5"
-        testInstrumentationRunner "com.google.android.apps.common.testing.testrunner.GoogleInstrumentationTestRunner"
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -27,6 +26,10 @@ android {
         androidTest {
             setRoot('src/test')
         }
+    }
+
+    packagingOptions {
+        exclude 'LICENSE.txt'
     }
 
     applicationVariants.all { variant ->
@@ -49,48 +52,25 @@ android {
     }
 }
 
-robolectric {
-    include '**/*Test.class'
-    exclude '**/espresso/**/*.class'
-}
-
 dependencies {
-
     compile 'com.android.support:appcompat-v7:21.0.+'
     compile 'com.google.android.gms:play-services:6.1.11'
     compile 'com.makeramen:roundedimageview:1.2.+'
     compile files('libs/android-query.0.26.7.jar')
 
+    compile 'com.android.support:support-annotations:21.0.3'
+
     // Espresso
-    androidTestCompile files('libs/espresso-1.1.jar', 'libs/testrunner-1.1.jar', 'libs/testrunner-runtime-1.1.jar')
-    androidTestCompile 'com.google.guava:guava:14.0.1'
-    androidTestCompile 'com.squareup.dagger:dagger:1.1.0'
-    androidTestCompile 'org.hamcrest:hamcrest-integration:1.1'
-    androidTestCompile 'org.hamcrest:hamcrest-core:1.1'
-    androidTestCompile 'org.hamcrest:hamcrest-library:1.1'
+    androidTestCompile 'com.android.support.test.espresso:espresso-core:2.0'
+    androidTestCompile 'com.android.support.test:testing-support-lib:0.1'
 
-    androidTestCompile('junit:junit:4.+') {
-        exclude module: 'hamcrest-core'
-    }
+    // Mockito
+    androidTestCompile 'org.mockito:mockito-core:1.9.5'
+    androidTestCompile 'com.google.dexmaker:dexmaker-mockito:1.1'
+    androidTestCompile 'com.google.dexmaker:dexmaker:1.1'
 
-    androidTestCompile files('libs/robolectric-3.0-20150113.000047-99-jar-with-dependencies.jar')
-
-//    androidTestCompile('org.robolectric:robolectric:3.0-SNAPSHOT') {
-//        exclude module: 'classworlds'
-//        exclude module: 'commons-logging'
-//        exclude module: 'httpclient'
-//        exclude module: 'maven-artifact'
-//        exclude module: 'maven-artifact-manager'
-//        exclude module: 'maven-error-diagnostics'
-//        exclude module: 'maven-model'
-//        exclude module: 'maven-project'
-//        exclude module: 'maven-settings'
-//        exclude module: 'plexus-container-default'
-//        exclude module: 'plexus-interpolation'
-//        exclude module: 'plexus-utils'
-//        exclude module: 'wagon-file'
-//        exclude module: 'wagon-http-lightweight'
-//        exclude module: 'wagon-provider-api'
-//    }
+    // Dagger
+    compile 'com.squareup.dagger:dagger:1.2.2'
+    provided 'com.squareup.dagger:dagger-compiler:1.2.2'
 
 }
