+++ App/src/main/java/com/fsecure/lokki/utils/ContactUtils.java
@@ -1,0 +1,11 @@
+package com.fsecure.lokki.utils;
+
+import android.content.Context;
+
+import org.json.JSONObject;
+
+public interface ContactUtils {
+
+    public JSONObject listContacts(Context context);
+
+}
