+++ App/src/test/java/com/fsecure/lokki/espresso/WelcomeScreensTest.java
@@ -1,0 +1,59 @@
+package com.fsecure.lokki.espresso;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.test.ActivityInstrumentationTestCase2;
+import android.test.suitebuilder.annotation.LargeTest;
+
+import com.fsecure.lokki.MainActivity;
+import com.fsecure.lokki.R;
+
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
+
+@LargeTest
+public class WelcomeScreensTest extends ActivityInstrumentationTestCase2<MainActivity> {
+
+    public WelcomeScreensTest() {
+        super(MainActivity.class);
+    }
+
+    @Override
+    public void setUp() throws Exception {
+        super.setUp();
+        clearAppData();
+        getActivity();
+    }
+
+    @Override
+    public void tearDown() throws Exception {
+        getActivity().firstTimeLaunch = null;
+        super.tearDown();
+    }
+
+    public void testWelcomeTextIsOnScreen() {
+        onView(withText(R.string.welcome_title)).check(matches(isDisplayed()));
+    }
+
+    public void testContinueButtonTakesToTermsScreen() {
+        onView(withText(R.string.continue_with_terms)).perform(click());
+        onView(withText(R.string.terms_title)).check(matches(isDisplayed()));
+    }
+
+    public void testAgreeOnTermsTakesToRegistrationScreen() {
+        onView(withText(R.string.continue_with_terms)).perform(click());
+        onView(withText(R.string.i_agree)).perform(click());
+        onView(withText(R.string.signup_explanation)).check(matches(isDisplayed()));
+    }
+
+    private void clearAppData() {
+        Context context = getInstrumentation().getTargetContext();
+        SharedPreferences.Editor editor = context.getSharedPreferences(context.getPackageName(), Context.MODE_PRIVATE).edit();
+        editor.clear();
+        editor.commit();
+    }
+
+}
