+++ build.gradle
@@ -7,15 +7,11 @@ buildscript {
 
     dependencies {
         classpath 'com.android.tools.build:gradle:1.0.1'
-        classpath 'org.robolectric:robolectric-gradle-plugin:0.14.+'
     }
 }
 
 allprojects {
     repositories {
         mavenCentral()
-        maven {
-            url "https://oss.sonatype.org/content/repositories/snapshots"
-        }
     }
 }
